+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix detection of Firefox on iPad.
 - Fix detection of Linux ARM-based Android.
 - Add detection of Chromium Edge on Windows.
+- Add detection of OkHttp.
 
 ## 0.5.2